Set Portrait Orientation to Whole APP Set AppBar title to corresponding Story Title
