+++ app/src/main/java/com/gujja/ajay/fourthver/activities/Adult_Group.java
@@ -16,42 +16,44 @@
 
 */
 
-package com.gujja.ajay.fourthver;
+package com.gujja.ajay.fourthver.activities;
 
 import android.os.Bundle;
 import android.view.View;
 import android.widget.ImageButton;
 
-import com.gujja.ajay.fourthver.Adapters.Child_Age_Adapter;
+import com.gujja.ajay.fourthver.Adapters.Adult_Age_Adapter;
+import com.gujja.ajay.fourthver.R;
 import com.tbuonomo.viewpagerdotsindicator.DotsIndicator;
 
 import androidx.appcompat.app.AppCompatActivity;
+import androidx.viewpager.widget.ViewPager;
 import butterknife.BindView;
 import butterknife.ButterKnife;
 
-public class Child_Group extends AppCompatActivity implements androidx.viewpager.widget.ViewPager.OnPageChangeListener, View.OnClickListener {
+public class Adult_Group extends AppCompatActivity implements ViewPager.OnPageChangeListener, View.OnClickListener {
 
-    @BindView(R.id.ViewPager)
-    androidx.viewpager.widget.ViewPager ViewPager;
-    @BindView(R.id.layout_dots)
+    @BindView(R.id.Adult_Group_ViewPager)
+    ViewPager AdultGroupViewPager;
+    @BindView(R.id.Adult_Group_dots)
     DotsIndicator dotsIndicator;
-    @BindView(R.id.Child_arrow)
-    ImageButton ChildArrow;
+    @BindView(R.id.Adult_arrow)
+    ImageButton Adult_Arrow;
 
     int currentPage;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_child_group);
+        setContentView(R.layout.activity_adult__group);
         ButterKnife.bind(this);
 
-        Child_Age_Adapter my_adapter = new Child_Age_Adapter(this);
-        ViewPager.setAdapter(my_adapter);
-        dotsIndicator.setViewPager(ViewPager);
+        Adult_Age_Adapter adultAgeAdapter = new Adult_Age_Adapter(this);
+        AdultGroupViewPager.setAdapter(adultAgeAdapter);
+        dotsIndicator.setViewPager(AdultGroupViewPager);
 
-        ViewPager.addOnPageChangeListener(this);
-        ChildArrow.setOnClickListener(this);
+        AdultGroupViewPager.addOnPageChangeListener(this);
+        Adult_Arrow.setOnClickListener(this);
     }
 
     @Override
@@ -62,11 +64,11 @@ public class Child_Group extends AppCompatActivity implements androidx.viewpager
     public void onPageSelected(int position) {
         currentPage = position;
         if (position == 2) {
-            ChildArrow.setEnabled(false);
-            ChildArrow.setVisibility(View.INVISIBLE);
+            Adult_Arrow.setEnabled(false);
+            Adult_Arrow.setVisibility(View.INVISIBLE);
         } else {
-            ChildArrow.setEnabled(true);
-            ChildArrow.setVisibility(View.VISIBLE);
+            Adult_Arrow.setEnabled(true);
+            Adult_Arrow.setVisibility(View.VISIBLE);
         }
     }
 
@@ -76,6 +78,6 @@ public class Child_Group extends AppCompatActivity implements androidx.viewpager
 
     @Override
     public void onClick(View view) {
-        ViewPager.setCurrentItem(currentPage + 1);
+        AdultGroupViewPager.setCurrentItem(currentPage + 1);
     }
 }
