+++ app/src/main/java/com/gujja/ajay/fourthver/activities/Child_Group.java
@@ -16,43 +16,43 @@
 
 */
 
-package com.gujja.ajay.fourthver;
+package com.gujja.ajay.fourthver.activities;
 
 import android.os.Bundle;
 import android.view.View;
 import android.widget.ImageButton;
 
-import com.gujja.ajay.fourthver.Adapters.Adult_Age_Adapter;
+import com.gujja.ajay.fourthver.Adapters.Child_Age_Adapter;
+import com.gujja.ajay.fourthver.R;
 import com.tbuonomo.viewpagerdotsindicator.DotsIndicator;
 
 import androidx.appcompat.app.AppCompatActivity;
-import androidx.viewpager.widget.ViewPager;
 import butterknife.BindView;
 import butterknife.ButterKnife;
 
-public class Adult_Group extends AppCompatActivity implements ViewPager.OnPageChangeListener, View.OnClickListener {
+public class Child_Group extends AppCompatActivity implements androidx.viewpager.widget.ViewPager.OnPageChangeListener, View.OnClickListener {
 
-    @BindView(R.id.Adult_Group_ViewPager)
-    ViewPager AdultGroupViewPager;
-    @BindView(R.id.Adult_Group_dots)
+    @BindView(R.id.ViewPager)
+    androidx.viewpager.widget.ViewPager ViewPager;
+    @BindView(R.id.layout_dots)
     DotsIndicator dotsIndicator;
-    @BindView(R.id.Adult_arrow)
-    ImageButton Adult_Arrow;
+    @BindView(R.id.Child_arrow)
+    ImageButton ChildArrow;
 
     int currentPage;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_adult__group);
+        setContentView(R.layout.activity_child_group);
         ButterKnife.bind(this);
 
-        Adult_Age_Adapter adultAgeAdapter = new Adult_Age_Adapter(this);
-        AdultGroupViewPager.setAdapter(adultAgeAdapter);
-        dotsIndicator.setViewPager(AdultGroupViewPager);
+        Child_Age_Adapter my_adapter = new Child_Age_Adapter(this);
+        ViewPager.setAdapter(my_adapter);
+        dotsIndicator.setViewPager(ViewPager);
 
-        AdultGroupViewPager.addOnPageChangeListener(this);
-        Adult_Arrow.setOnClickListener(this);
+        ViewPager.addOnPageChangeListener(this);
+        ChildArrow.setOnClickListener(this);
     }
 
     @Override
@@ -63,11 +63,11 @@ public class Adult_Group extends AppCompatActivity implements ViewPager.OnPageCh
     public void onPageSelected(int position) {
         currentPage = position;
         if (position == 2) {
-            Adult_Arrow.setEnabled(false);
-            Adult_Arrow.setVisibility(View.INVISIBLE);
+            ChildArrow.setEnabled(false);
+            ChildArrow.setVisibility(View.INVISIBLE);
         } else {
-            Adult_Arrow.setEnabled(true);
-            Adult_Arrow.setVisibility(View.VISIBLE);
+            ChildArrow.setEnabled(true);
+            ChildArrow.setVisibility(View.VISIBLE);
         }
     }
 
@@ -77,6 +77,6 @@ public class Adult_Group extends AppCompatActivity implements ViewPager.OnPageCh
 
     @Override
     public void onClick(View view) {
-        AdultGroupViewPager.setCurrentItem(currentPage + 1);
+        ViewPager.setCurrentItem(currentPage + 1);
     }
 }
