+++ app/src/main/java/com/gujja/ajay/fourthver/activities/Mid_Age_Group.java
@@ -16,13 +16,14 @@
 
 */
 
-package com.gujja.ajay.fourthver;
+package com.gujja.ajay.fourthver.activities;
 
 import android.os.Bundle;
 import android.view.View;
 import android.widget.ImageButton;
 
-import com.gujja.ajay.fourthver.Adapters.AgeSelectionAdapter;
+import com.gujja.ajay.fourthver.Adapters.Mid_Age_Adapter;
+import com.gujja.ajay.fourthver.R;
 import com.tbuonomo.viewpagerdotsindicator.DotsIndicator;
 
 import androidx.appcompat.app.AppCompatActivity;
@@ -30,58 +31,57 @@ import androidx.viewpager.widget.ViewPager;
 import butterknife.BindView;
 import butterknife.ButterKnife;
 
-public class AgeSelection extends AppCompatActivity implements View.OnClickListener, ViewPager.OnPageChangeListener {
+public class Mid_Age_Group extends AppCompatActivity implements ViewPager.OnPageChangeListener, View.OnClickListener {
 
-    @BindView(R.id.Age_selection_ViewPager)
-    ViewPager Age_ViewPager;
-    @BindView(R.id.Age_selection_layout_dots)
-    DotsIndicator dotsIndicator;
-    @BindView(R.id.Age_selection_arrow)
-    ImageButton Age_Arrow;
+    @BindView(R.id.Mid_Age_ViewPager)
+    ViewPager MidAge_ViewPager;
+    @BindView(R.id.Mid_Age_layout_dots)
+    DotsIndicator MidAge_Layout_Dots;
+    @BindView(R.id.Mid_Arrow)
+    ImageButton MidArrow;
 
     int currentPage;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_age_selection);
+        setContentView(R.layout.activity_mid__age__group);
         ButterKnife.bind(this);
 
-        AgeSelectionAdapter age_selection_adapter = new AgeSelectionAdapter(this);
-        Age_ViewPager.setAdapter(age_selection_adapter);
-        dotsIndicator.setViewPager(Age_ViewPager);
+        Mid_Age_Adapter mid_age_adapter = new Mid_Age_Adapter(this);
+        MidAge_ViewPager.setAdapter(mid_age_adapter);
 
-        Age_ViewPager.addOnPageChangeListener(this);
-        Age_Arrow.setOnClickListener(this);
+        MidAge_Layout_Dots.setViewPager(MidAge_ViewPager);
 
+        MidAge_ViewPager.addOnPageChangeListener(this);
+        MidArrow.setOnClickListener(this);
     }
 
-    @Override
-    public void onClick(View view) {
-        Age_ViewPager.setCurrentItem(currentPage + 1);
-
-    }
 
     @Override
     public void onPageScrolled(int position, float positionOffset, int positionOffsetPixels) {
+
     }
 
     @Override
     public void onPageSelected(int position) {
         currentPage = position;
         if (position == 2) {
-            Age_Arrow.setEnabled(false);
-            Age_Arrow.setVisibility(View.INVISIBLE);
+            MidArrow.setEnabled(false);
+            MidArrow.setVisibility(View.INVISIBLE);
         } else {
-            Age_Arrow.setEnabled(true);
-            Age_Arrow.setVisibility(View.VISIBLE);
+            MidArrow.setEnabled(true);
+            MidArrow.setVisibility(View.VISIBLE);
         }
     }
 
     @Override
     public void onPageScrollStateChanged(int state) {
+
+    }
+
+    @Override
+    public void onClick(View view) {
+        MidAge_ViewPager.setCurrentItem(currentPage + 1);
     }
 }
-
-
-
