+++ app/src/main/java/com/gujja/ajay/fourthver/stories/Foolish_Donkey.java
@@ -40,8 +40,7 @@ import butterknife.BindView;
 import butterknife.ButterKnife;
 import pl.droidsonroids.gif.GifImageView;
 
-
-public class FourthStory extends AppCompatActivity implements TextToSpeech.OnInitListener, View.OnClickListener {
+public class Foolish_Donkey extends AppCompatActivity implements TextToSpeech.OnInitListener, View.OnClickListener {
 
     int i = 0;
     int j = 0;
@@ -49,50 +48,50 @@ public class FourthStory extends AppCompatActivity implements TextToSpeech.OnIni
     float pitch = 0.8f;
     private TextToSpeech tts;
 
-    @BindView(R.id.friend_textview)
-    TextView friendSentence;
-    @BindView(R.id.friend_TextWord)
-    TextView friendWord;
-    @BindView(R.id.friend_Button_speak)
-    ImageButton friendButtonSpeak;
-    @BindView(R.id.friend_button_stop)
-    ImageButton friendButtonStop;
-    @BindView(R.id.friend_SignImage)
-    GifImageView friendSignGifs;
+    @BindView(R.id.fool_donkey_textview)
+    TextView foolDonkeySentence;
+    @BindView(R.id.fool_donkey_TextWord)
+    TextView foolDonkeyWord;
+    @BindView(R.id.fool_donkey_Button_speak)
+    ImageButton foolDonkeyButtonSpeak;
+    @BindView(R.id.fool_donkey_button_stop)
+    ImageButton foolDonkeyButtonStop;
+    @BindView(R.id.fool_donkey_SignImage)
+    GifImageView foolDonkeySignGifs;
 
-    String[] friend_word = {
-            "Once", "upon", "a", "time", "there", "lived", "a", "lion", "in", "a", "forest", " ",
-            "One", "day", "after", "a", "heavy", "meal", " ",
-            "It", "was", "sleeping", "under", "a", "tree", " ",
-            "After", "a", "while", "there", "came", "a", "mouse", "and", "it", "started", "to", "play", "on", "the", "lion", " ",
-            "Suddenly", "the", "lion", "got", "up", "with", "anger", "and", "looked", "for", "those", "who", "disturbed", "its", "nice", "sleep", " ",
-            "Then", "it", "saw", "a", "small", "mouse", "standing", "trembling", "with", "fear", " ",
-            "The", "lion", "jumped", "on", "it", "and", "started", "to", "kill", "it", " ",
-            "The", "mouse", "requested", "the", "lion", "to", "forgive", "it", " ",
-            "The", "lion", "felt", "pity", "and", "left", "it", " ",
-            "The", "mouse", "ran", "away", " ",
-            "On", "another", "day", "the", "lion", "was", "caught", "in", "a", "net", "by", "a", "hunter", " ",
-            "The", "mouse", "came", "there", "and", "cut", "the", "net", " ",
-            "Thus", "it", "escaped", " ",
-            "There", "after", "the", "mouse", "and", "the", "lion", "became", "friends", " ",
-            "They", "lived", "happily", "in", "the", "forest", "afterwards"
+
+    String[] foolish = {
+            "A", "salt", "seller", "used", "to", "carry", "the", "salt", "bag", "on", "his",
+            "donkey", "to", "the", "market", "every", "day", " ",
+            "On", "the", "way", "they", "had", "to", "cross", "a", "stream", " ",
+            "One", "day", "the", "donkey", "suddenly", "tumbled", "down", "the", "stream",
+            "and", "the", "salt", "bag", "also", "fell", "into", "the", "water", " ",
+            "The", "salt", "dissolved", "in", "the", "water", "and", "hence", "the", "bag", "became", "very", "light", "to", "carry", " ",
+            "The", "donkey", "was", "happy", " ",
+            "Then", "the", "donkey", "started", "to", "play", "the", "same", "trick", "every", "day", " ",
+            "The", "salt", "seller", "came", "to", "understand", "the", "trick", "and", "decided", "to", "teach", "a", "lesson", "to", "it", " ",
+            "The", "next", "day", "he", "loaded", "a", "cotton", "bag", "on", "the", "donkey", " ",
+            "Again", "it", "played", "the", "same", "trick", "hoping", "that", "the", "cotton", "bag", "would", "be", "still", "become", "lighter", " ",
+            "But", "the", "dampened", "cotton", "became", "very", "heavy", "to", "carry", "and", "the", "donkey", "suffered", " ",
+            "It", "learnt", "a", "lesson", " ",
+            "It", "didn't", "play", "the", "trick", "anymore", "after", "that", "day", "and", "the", "seller", "was", "happy"
     };
-    String[] friend_sentence = {
-            "Once upon a time there lived a lion in a forest.",
-            "One day after a heavy meal.",
-            "It was sleeping under a tree.",
-            "After a while, there came a mouse and it started to play on the lion.",
-            "Suddenly the lion got up with anger and looked for those who disturbed its nice sleep.",
-            "Then it saw a small mouse standing trembling with fear.",
-            "The lion jumped on it and started to kill it.",
-            "The mouse requested the lion to forgive it.",
-            "The lion felt pity and left it.",
-            "The mouse ran away.",
-            "On another day, the lion was caught in a net by a hunter.",
-            "The mouse came there and cut the net.",
-            "Thus it escaped.",
-            "There after, the mouse and the lion became friends.",
-            "They lived happily in the forest afterwards."};
+
+    String[] donkey = {
+            "A salt seller used to carry the salt bag on his donkey to the market every day.\n",
+            "On the way they had to cross a stream.\n",
+            "One day the donkey suddenly tumbled down the stream and the salt bag also fell into the water.\n",
+            "The salt dissolved in the water and hence the bag became very light to carry.\n",
+            "The donkey was happy.\n",
+            "Then the donkey started to play the same trick every day.\n",
+            "The salt seller came to understand the trick and decided to teach a lesson to it.\n",
+            "The next day he loaded a cotton bag on the donkey.\n",
+            "Again it played the same trick hoping that the cotton bag would be still become lighter.\n",
+            "But the dampened cotton became very heavy to carry and the donkey suffered.\n",
+            "It learnt a lesson.\n",
+            "It didn’t play the trick anymore after that day, and the seller was happy."
+    };
+
     String[] stopwords = {
             "the", "all", "into", "loaf", "but", "for", "and", "at", "found", "of", "in", "squeezed", "hole", "to",
             "have", "caught", "gave", "came", "on", "become", "trick", "with", "carry", "cotton", "that", "felt",
@@ -106,21 +105,23 @@ public class FourthStory extends AppCompatActivity implements TextToSpeech.OnIni
             "capsized","speechless","grief","what","cheating","dealings","supreme","anymore","seller","won't","favor","about","instead"
     };
 
-    //Initialization of Utterance Listener
+
+
     UtteranceProgressListener mProgressListener = new UtteranceProgressListener() {
         @Override
-        public void onStart(final String utteranceId) {
+        public void onStart(String utteranceId) {
 
             // For Highlighting Spoken Words
-            String Replce = "<span style= 'background-color:green'>" + utteranceId + "</span>";
-            friendWord.setText(Html.fromHtml(Replce));
+            String Replace = "<span style= 'background-color:green'>" + utteranceId + "</span>";
+            foolDonkeyWord.setText(Html.fromHtml(Replace));
+
 
             if (utteranceId.toLowerCase(Locale.getDefault()).equals("try") || utteranceId.toLowerCase(Locale.getDefault()).equals("catch") || utteranceId.toLowerCase(Locale.ROOT).equals("while")) {
                 int gif_view = getResources().getIdentifier(utteranceId.toLowerCase(Locale.getDefault()) + "1", "raw", getPackageName());
-                friendSignGifs.setImageResource(gif_view);
+                foolDonkeySignGifs.setImageResource(gif_view);
             } else {
                 int gif_view = getResources().getIdentifier(utteranceId.toLowerCase(Locale.getDefault()), "raw", getPackageName());
-                friendSignGifs.setImageResource(gif_view);
+                foolDonkeySignGifs.setImageResource(gif_view);
             }
 
         }
@@ -134,14 +135,13 @@ public class FourthStory extends AppCompatActivity implements TextToSpeech.OnIni
                     @Override
                     public void run() {
                         j++;
-                        friendSentence.setText(friend_sentence[j]);
+                        foolDonkeySentence.setText(donkey[j]);
                     }
-                };
-
+                }.start();
             }
-            // For Incrementing Words
+
             i = i + 1;
-            speak(friend_word, i);
+            speak(foolish, i);
 
         }
 
@@ -154,40 +154,40 @@ public class FourthStory extends AppCompatActivity implements TextToSpeech.OnIni
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_fourth_story);
+        setContentView(R.layout.activity_third_story);
         ButterKnife.bind(this);
+        Foolish_Donkey.this.setTitle("Foolish Donkey");
 
-        //Setting Default Text
-        friendSentence.setText(friend_sentence[0]);
-        friendWord.setText(friend_word[0]);
+        foolDonkeySentence.setText(donkey[0]);
+        foolDonkeyWord.setText(foolish[0]);
 
         // Initialization of Text To Speech
-        tts = new TextToSpeech(FourthStory.this, this);
+        tts = new TextToSpeech(Foolish_Donkey.this, this);
 
         // Tracking of Words
         tts.setOnUtteranceProgressListener(mProgressListener);
 
-        friendButtonSpeak.setOnClickListener(this);
-        friendButtonStop.setOnClickListener(this);
+        /* @OnClick Listener */
+        foolDonkeyButtonSpeak.setOnClickListener(this);
+        foolDonkeyButtonStop.setOnClickListener(this);
+
     }
 
     private void stop() {
-
         tts.stop();
         tts.shutdown();
-        friendSentence.setText(friend_sentence[0]);
-        friendWord.setText(friend_word[0]);
-        friendSignGifs.setVisibility(View.INVISIBLE);
+        foolDonkeySentence.setText(donkey[0]);
+        foolDonkeyWord.setText(foolish[0]);
+        foolDonkeySignGifs.setVisibility(View.INVISIBLE);
     }
 
     private void speak(String[] text, int i) {
-
         tts.setSpeechRate(speed);
         tts.setPitch(pitch);
         HashMap<String, String> map = new HashMap<>();
 
         for (String stopword : stopwords) {
-            if (friend_word[i].toLowerCase(Locale.getDefault()).equals(stopword)) {
+            if (foolish[i].toLowerCase(Locale.getDefault()).equals(stopword)) {
                 char[] alphabet_array = stopword.toCharArray();
 
                 for (char c : alphabet_array) {
@@ -195,19 +195,20 @@ public class FourthStory extends AppCompatActivity implements TextToSpeech.OnIni
                     tts.setSpeechRate(0.5f);
                     tts.speak(String.valueOf(c), TextToSpeech.QUEUE_ADD, map);
                 }
+
             }
         }
 
-        map.put(TextToSpeech.Engine.KEY_PARAM_UTTERANCE_ID, friend_word[i]);
-        tts.speak(friend_word[i], TextToSpeech.QUEUE_ADD, map);
 
+        map.put(TextToSpeech.Engine.KEY_PARAM_UTTERANCE_ID, foolish[i]);
+        tts.speak(foolish[i], TextToSpeech.QUEUE_ADD, map);
     }
-
 
     @Override
     public boolean onCreateOptionsMenu(Menu menu) {
         MenuInflater inflater = getMenuInflater();
         inflater.inflate(R.menu.speedmeter, menu);
+
         return true;
     }
 
@@ -242,6 +243,7 @@ public class FourthStory extends AppCompatActivity implements TextToSpeech.OnIni
         stop();
         super.onBackPressed();
         this.finish();
+
     }
 
     @Override
@@ -255,18 +257,16 @@ public class FourthStory extends AppCompatActivity implements TextToSpeech.OnIni
     @Override
     public void onClick(View view) {
         switch (view.getId()){
-            case  R.id.friend_Button_speak:
-                friendButtonSpeak.setEnabled(false);
-                speak(friend_word,i);
+            case  R.id.fool_donkey_Button_speak:
+                foolDonkeyButtonSpeak.setEnabled(false);
+                speak(foolish,i);
                 break;
 
-            case  R.id.friend_button_stop:
-                friendButtonSpeak.setEnabled(true);
+            case  R.id.fool_donkey_button_stop:
+                foolDonkeyButtonSpeak.setEnabled(true);
                 stop();
                 break;
         }
     }
 
-    private abstract class runnable implements Runnable {
-    }
 }
