+++ app/src/main/res/layout/activity_adult__group.xml
@@ -25,7 +25,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Adult_Group">
+    tools:context=".activities.Adult_Group">
 
     <androidx.viewpager.widget.ViewPager
         android:foregroundGravity="center"
