+++ app/src/main/res/layout/activity_age_selection.xml
@@ -27,7 +27,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#fff"
-    tools:context=".AgeSelection">
+    tools:context=".activities.AgeSelection">
