+++ app/src/main/res/layout/activity_fourth_story.xml
@@ -23,7 +23,7 @@
     android:layout_width="match_parent"
     android:orientation="vertical"
     android:layout_height="match_parent"
-    tools:context=".stories.FourthStory">
+    tools:context=".stories.Lion_Mouse">
 
     <LinearLayout
         android:layout_width="match_parent"
