+++ app/src/main/res/layout/activity_mid__age__group.xml
@@ -23,7 +23,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Mid_Age_Group">
+    tools:context=".activities.Mid_Age_Group">
 
 
     <androidx.viewpager.widget.ViewPager
