+++ app/src/main/res/layout/activity_third_story.xml
@@ -23,7 +23,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".stories.ThirdStory">
+    tools:context=".stories.Foolish_Donkey">
 
 
     <LinearLayout
